refactor: CategorySelectionFragment 패키지 이동
